Correct time averaging of flux variables, simplify code
This commit corrects the time averaging of flux variables, which previously
was not including the Jan 1 entry of the following calendar year.  That entry
technically belongs to the previous year for flux calculation.  This commit does
this by revamping the way time is handled in the flux time averaging.  This also
corrects the calculation of time bounds.

This commit also:
* stops calling the translate_GL_and_calving_flux_edge2cell routine, which we no
longer want to use.  The GL flux is now calculated correctly in MALI.  The calving
flux calculation should also be moved to MALI, but has not been implemented yet, because
we are not running with calving yet.  Eliminating the call to this function also allows
us to eliminate some ncks and shutil steps, because they are no longer needed.  That
also speeds things up, because some of those were kind of slow.  The code here is updated
to use the new MALI GL flux variable on cells.
* Enables dH/dt calculation.

+++ landice/output_processing_li/ismip6_postprocessing/post_process_mali_to_ismip6.py
@@ -134,30 +134,16 @@ def main():
         # call the function that adds and renames state vars as requested by the
         # ISMIP6 protocol
         print(  "copying flux input file")
-        input_fname = os.path.basename(args.input_file_flux)
-        input_file_copy = f"copy_{input_fname}"
-        shutil.copy(args.input_file_flux, input_file_copy)
-
-        # append grid cell data to the flux data
-        print(  "appending grid cell data to flux input file")
-        command = ["ncks", "-A", "-v",
-                   "cellsOnEdge,nEdgesOnCell,edgesOnCell,dvEdge,areaCell",
-                   args.input_file_init, input_file_copy]
-        check_call(command)
 
         # take time (yearly) average for the flux variables
         tmp_file1 = "flux_time_avg.nc"
-        tmp_file2 = "flux_on_cell.nc"
-        do_time_avg_flux_vars(input_file_copy, tmp_file1)
-
-        # translate grounding line and calving fluxes from cell edges to center
-        translate_GL_and_calving_flux_edge2cell(tmp_file1, tmp_file2)
+        do_time_avg_flux_vars(args.input_file_flux, tmp_file1)
 
         # remap data from the MALI unstructured mesh to the ISMIP6 P-S grid
         processed_file_flux = f'processed_' \
                               f'{os.path.basename(args.input_file_flux)}'
         command = ["ncremap",
-                   "-i", tmp_file2,
+                   "-i", tmp_file1,
                    "-o", processed_file_flux,
                    "-m", mapping_file,
                    "-P", "mpas"]
@@ -168,10 +154,10 @@ def main():
                                      args.ismip6_grid_file,
                                      args.exp, output_path)
 
-        os.remove(tmp_file1)
-        os.remove(tmp_file2)
-        os.remove(processed_file_flux)
-        os.remove(input_file_copy)
+        cleanUp = True
+        if cleanUp:
+            os.remove(tmp_file1)
+            os.remove(processed_file_flux)
         print("---Processing flux file complete---\n")
     print("---All processing complete---")
